final adjustments

+++ src/main/java/drai/dev/stackthecards/data/CardConnectionEntry.java
@@ -70,21 +70,21 @@ public class CardConnectionEntry {
         }
         if(json.containsKey(JSON_X_MODIFIER_KEY)){
             try{
-                connectionEntry.xModifier = (float)(double) json.get(JSON_X_MODIFIER_KEY);
+                connectionEntry.xModifier = (float)(long) json.get(JSON_X_MODIFIER_KEY);
             } catch (Exception e){
                 throw new MalformedJsonException("Text format isItalic was malformed: "+e.getMessage());
             }
         }
         if(json.containsKey(JSON_Y_MODIFIER_KEY)){
             try{
-                connectionEntry.yModifier = (float)(double) json.get(JSON_Y_MODIFIER_KEY);
+                connectionEntry.yModifier = (float)(long) json.get(JSON_Y_MODIFIER_KEY);
             } catch (Exception e){
                 throw new MalformedJsonException("Text format isItalic was malformed: "+e.getMessage());
             }
         }
         if(json.containsKey(JSON_LAYER_KEY)) {
             try {
-                connectionEntry.layer = (int) json.get(JSON_LAYER_KEY);
+                connectionEntry.layer = (int)(long) json.get(JSON_LAYER_KEY);
             } catch (Exception e) {
                 throw new MalformedJsonException("Text format isItalic was malformed: "+e.getMessage());
             }
